TRint & Socket
git-svn-id: svn+ssh://savannah.psi.ch/afs/psi.ch/project/meg/svn/rome/trunk@89 306b0a18-5003-0410-92f3-cc1c673a7936

+++ src/ROMETask.cpp
@@ -77,9 +77,9 @@ char* ROMETask::GetTime()
    // Returns the elapsed time in a readable format
    int runTime = (int)fWatch.RealTime();
    int milli = (int)((fWatch.RealTime()-runTime)*1000);
-   fTimeString.SetFormatted("%i%i:%i%i:%i%i:%d", runTime / 36000, (runTime % 36000) / 3600,
-           (runTime % 3600) / 600, (runTime % 600) / 60, (runTime % 60) / 10,
-           runTime % 10,milli);
+   fTimeString.SetFormatted("%d%d:%d%d:%d%d:%d", (int)(runTime / 36000), (int)((runTime % 36000) / 3600),
+           (int)((runTime % 3600) / 600), (int)((runTime % 600) / 60), (int)((runTime % 60) / 10),
+           (int)(runTime % 10),milli);
    return (char*)fTimeString.Data();
 }
